+++ app/controllers/recording_studio_plugin_sdk_template/home_controller.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module RecordingStudioPluginSdkTemplate
-  class HomeController < ApplicationController
-    def index
-      head :ok
-    end
-  end
-end