+++ config/routes.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
 
 RecordingStudioPluginSdkTemplate::Engine.routes.draw do
-  root "home#index"
 end