+++ lib/generators/recording_studio_plugin_sdk_template/install/templates/recording_studio_plugin_sdk_template_initializer.rb
@@ -1,12 +1,3 @@
 # frozen_string_literal: true
 
-RecordingStudioPluginSdkTemplate.configure do |config|
-  # Set your API key (recommended to use ENV or Rails credentials)
-  # config.api_key = ENV["RECORDING_STUDIO_PLUGIN_SDK_TEMPLATE_API_KEY"]
-
-  # Enable optional feature X
-  # config.enable_feature_x = false
-
-  # Timeout in seconds for external calls
-  # config.timeout = 5
-end
+RecordingStudioPluginSdkTemplate.configure