+++ test/dummy/config/initializers/recording_studio_plugin_sdk_template.rb
@@ -1,5 +1,3 @@
 # frozen_string_literal: true
 
-GemTemplate.configure do |config|
-  # config.api_key = "..."
-end
+RecordingStudioPluginSdkTemplate.configure