+++ test/rename_gem_identity_test.rb
@@ -6,8 +6,10 @@ class RenameGemIdentityTest < Minitest::Test
   def test_rename_script_rewrites_leftover_template_homepages
     script = File.read(File.expand_path("../bin/rename_gem", __dir__))
 
-    assert_includes script, "https://github.com/bowerbird-app/recording_studio_plugin_sdk_template"
-    assert_includes script, "https://github.com/bowerbird-app/recording_studio_plugin_sdk_template"
+    assert_includes script, "https://github.com/bowerbird-app/RecordingStudio_gem_template"
+    assert_includes script, "https://github.com/bowerbird-app/gem_template"
+    assert_includes script, "def github_homepage(name)"
+    assert_includes script, "RecordingStudio_"
     assert_includes script, "rewrite_leftover_homepages!"
     assert_includes script, "leftover_template_identity?"
     assert_includes script, "README.md"